V1.1: Added server passwords, per-packet "verification", and ability to use shorts as packet IDs instead of strings!
Basically the title, the server can now have passwords. Instead of the app ID and version being sent with each packet, you can now choose whether to send those on each packet, send the hash codes of the strings, or send nothing. Another measure to reduce packet size is the ability to use shorts as the packet IDs instead of strings. (2 bytes compared to infinite number)