Actually fixed packet bug, few fixes and improvements
When trying the get the second parameter of the ServerReceive attribute, I accidentally indexed 0 instead of 1. That was fixed, as well as a bug when stopping and restarting the server (re-initialized the clients list). I also made the packet handler collecting log both the string and short ID, and finally called the ServerClose event before disconnecting the clients, in case you want to send 1 last packet (in my case, changing the scene back to the main menu)